Move HlslSourceRewriter to shared file
